requirements.txt
tensorflow==2.15.0
Pillow==9.1.0
numpy==1.23.5
scikit-learn==1.2.2  # For train/test split and SMOTE
imbalanced-learn # For oversampling
matplotlib
Flask==2.1.2
Werkzeug==2.0.3
Jinja2==3.1.2
gunicorn==20.1.0  # For production deployment (if needed)



